+++ odometrie.h
@@ -8,9 +8,9 @@ void init_odometrie();
 
 void actualise_position();
 
-int delta_mm(long int nbr_tick_D, long int nbr_tick_G);
-int alpha_millirad(long int nbr_tick_D, long int nbr_tick_G);
-void actualise_xy(int d_delta, int d_alpha, int theta, double * x, double * y);
+long int delta_mm(long int nbr_tick_D, long int nbr_tick_G);
+long int alpha_millirad(long int nbr_tick_D, long int nbr_tick_G);
+void actualise_xy(int d_delta, int d_alpha, int theta, float * x, float * y);
 int borne_angle(int angle);
 
 int get_delta_actuel();
